Extract constant behaviour to a module
