+++ lib/toy/behaviours.rb
@@ -25,5 +25,36 @@ module Toy
         ivar_map.delete(name.to_sym)
       end
     end
+
+    module Constants
+      def inspect
+        "#<#{self.class}>"
+      end
+
+      # Module instance methods
+      def constant_map
+        @constant_map ||= {}
+      end
+
+      def const_get(name)
+        name = name.to_sym
+        ::Kernel.raise ::NameError, "uninitialized constant #{inspect}::#{name}" unless constant_map.key?(name)
+
+        constant_map[name]
+      end
+
+      def const_set(name, value)
+        name = name.to_sym
+        ::Kernel.raise ::NameError unless name.match?(/^[A-Z][a-zA-Z_]*$/)
+
+        ::Kernel.warn "already initialized constant #{inspect}::#{name}" if constant_map.key?(name)
+
+        constant_map[name] = value
+      end
+
+      def constants
+        constant_map.keys.map(&:to_sym)
+      end
+    end
   end
 end
